+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### New Features
 
+- ghidra: experimental capa explorer extension (MVP) @vaishakh787
 - ghidra: support PyGhidra @mike-hunhoff #2788
 - vmray: extract number features from whitelisted void_ptr parameters (hKey, hKeyRoot) @adeboyedn #2835